feat: added index file for utils folder and updated the import of utils
