+++ src/algorithms/BubbleSort.ts
@@ -1,8 +1,4 @@
-import Bar from '../utils/Bar'
-import { IAlgorithm } from '../utils/interface'
-import changeStatusOfElement from '../utils/changeStatusOfElement'
-import swapElements from '../utils/swapElements'
-import { BarStatus } from '../utils/enum'
+import { IAlgorithm, changeStatusOfElement, swapElements, Bar, BarStatus } from '../utils'
 
 const { ACTIVE, SORTED, UNSORTED } = BarStatus
