+++ src/pages/Home/components/LegendContainer.tsx
@@ -5,7 +5,7 @@ import Legend from './Legend'
 import { Header } from '../../../shared'
 
 // utils
-import { BarStatus } from 'src/utils/enum'
+import { BarStatus } from '../../../utils'
 
 interface IProps {
 	legends: BarStatus[]
