+++ src/utils/index.ts
@@ -1,0 +1,6 @@
+export * from './Bar';
+export { default as changeStatusOfElement } from './changeStatusOfElement';
+export * from './interface';
+export * from './randomArrayGenerator';
+export { default as Steps } from './Steps';
+export { default as swapElements } from './swapElements';
